trail3/requirements.txt
pymongo
streamlit
langgraph
openai
google-generativeai
pandas
matplotlib
pydantic
requests
geopy
folium
streamlit-folium
groq
plotly
reportlab
python-dotenv
